Move files and make pseudo code less pseudo
