+++ src/ccpa/core.ts
@@ -1,0 +1,1 @@
+export const onIabConsentNotification = () => {};
