+++ src/ccpa/sourcepoint.ts
@@ -47,6 +47,3 @@ export const init = () => {
 
     document.body.appendChild(ccpaLib);
 };
-
-export const onGuConsentNotification = () => {};
-export const onIabConsentNotification = () => {};
